Merge Tile Entity Migration From Blockstate

+++ src/main/java/com/rabbitminers/extendedbogeys/mixin/MixinStandardBogeyBlock.java
@@ -1,28 +1,21 @@
 package com.rabbitminers.extendedbogeys.mixin;
 
-import com.jozufozu.flywheel.api.MaterialManager;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.mojang.math.Vector3f;
 import com.rabbitminers.extendedbogeys.bogey.styles.BogeyStyles;
 import com.rabbitminers.extendedbogeys.bogey.styles.IBogeyStyle;
-import com.rabbitminers.extendedbogeys.bogey.unlinked.UnlinkedBogeyTileEntity;
-import com.rabbitminers.extendedbogeys.bogey.unlinked.UnlinkedStandardBogeyBlock;
-import com.rabbitminers.extendedbogeys.bogey.util.LanguageKey;
 import com.rabbitminers.extendedbogeys.index.ExtendedBogeysBlocks;
 import com.rabbitminers.extendedbogeys.mixin_interface.BlockStates;
-import com.rabbitminers.extendedbogeys.mixin_interface.ICarriageBogeyStyle;
+import com.rabbitminers.extendedbogeys.mixin_interface.IStyledStandardBogeyTileEntity;
+import com.rabbitminers.extendedbogeys.mixin_interface.IStyledStandardBogeyBlock;
 import com.simibubi.create.AllBlocks;
-import com.simibubi.create.content.contraptions.components.structureMovement.bearing.MechanicalBearingBlock;
-import com.simibubi.create.content.contraptions.relays.elementary.CogwheelBlockItem;
 import com.simibubi.create.content.contraptions.relays.elementary.ShaftBlock;
 import com.simibubi.create.content.contraptions.wrench.WrenchItem;
-import com.simibubi.create.content.logistics.trains.entity.BogeyInstance;
-import com.simibubi.create.content.logistics.trains.entity.CarriageBogey;
 import com.simibubi.create.content.logistics.trains.track.StandardBogeyBlock;
 import com.simibubi.create.foundation.render.CachedBufferer;
 import com.simibubi.create.foundation.utility.Iterate;
-import net.minecraft.ChatFormatting;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.core.BlockPos;
@@ -33,13 +26,13 @@ import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.AirItem;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.*;
@@ -58,7 +51,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import java.util.EnumSet;
 
 @Mixin(StandardBogeyBlock.class)
-public abstract class MixinStandardBogeyBlock extends Block {
+public abstract class MixinStandardBogeyBlock extends Block implements IStyledStandardBogeyBlock {
     @Shadow @Final private boolean large;
     @Shadow public abstract EnumSet<Direction> getStickySurfaces(BlockGetter world, BlockPos pos, BlockState state);
     @Shadow @Final public static EnumProperty<Direction.Axis> AXIS;
@@ -66,31 +59,35 @@ public abstract class MixinStandardBogeyBlock extends Block {
     @Shadow protected abstract void renderBogey(float wheelAngle, PoseStack ms, int light, VertexConsumer vb, BlockState air);
 
     @Shadow protected abstract void renderLargeBogey(float wheelAngle, PoseStack ms, int light, VertexConsumer vb, BlockState air);
-
-    private static final Property<Integer> STYLE = BlockStates.STYLE;
-    private static final Property<Boolean> IS_FACING_FORWARD = BlockStates.IS_FACING_FOWARD;
-    private static final EnumProperty<DyeColor> PAINT_COLOUR = BlockStates.PAINT_COLOUR;
     public MixinStandardBogeyBlock(Properties pProperties) {
         super(pProperties);
-    }
-
-    @Inject(at = @At("HEAD"), method = "createBlockStateDefinition", remap = false)
-    public void createBlockStateDefenition(StateDefinition.Builder<Block, BlockState> builder, CallbackInfo ci) {
-        builder.add(STYLE);
-        builder.add(IS_FACING_FORWARD);
-
     }
 
     @Override
     public @NotNull InteractionResult use(BlockState state, Level level, BlockPos blockPos, Player player,
                                           InteractionHand interactionHand, BlockHitResult blockHitResult) {
+
+        BlockEntity be = level.getBlockEntity(blockPos);
+        IStyledStandardBogeyTileEntity te = (IStyledStandardBogeyTileEntity) be;
+        if (be == null)
+            return InteractionResult.FAIL;
+        CompoundTag tileData = be.getTileData();
+
         if (player.isShiftKeyDown() && !level.isClientSide && interactionHand == InteractionHand.MAIN_HAND
                 && player.getMainHandItem().getItem() == Items.AIR) {
-            BlockState unlinkedBlockState = large ? ExtendedBogeysBlocks.LARGE_UNLINKED_BOGEY.getDefaultState() : ExtendedBogeysBlocks.SMALL_UNLINKED_BOGEY.getDefaultState();
+            BlockState unlinkedBlockState = large
+                    ? ExtendedBogeysBlocks.LARGE_UNLINKED_BOGEY.getDefaultState()
+                    : ExtendedBogeysBlocks.SMALL_UNLINKED_BOGEY.getDefaultState();
 
             level.setBlock(blockPos, unlinkedBlockState
-                    .setValue(STYLE, state.getValue(STYLE))
                     .setValue(AXIS, state.getValue(AXIS)), 3);
+
+            IStyledStandardBogeyTileEntity usbte =
+                    (IStyledStandardBogeyTileEntity) level.getBlockEntity(blockPos);
+
+            CompoundTag unlinkedTileData = ((BlockEntity) usbte).getTileData();
+            usbte.setBogeyStyle(unlinkedTileData, te.getBogeyStyle(tileData));
+            usbte.setIsFacingForwards(unlinkedTileData, te.getIsFacingForwards(tileData));
 
             player.displayClientMessage(new TranslatableComponent("extendedbogeys.tooltips.unlink"), true);
 
@@ -99,8 +96,10 @@ public abstract class MixinStandardBogeyBlock extends Block {
 
         if (!player.isShiftKeyDown() && !level.isClientSide && interactionHand == InteractionHand.MAIN_HAND
                 && player.getMainHandItem().getItem() == Items.AIR) {
-            boolean facing = state.getValue(IS_FACING_FORWARD);
-            level.setBlock(blockPos, state.setValue(IS_FACING_FORWARD, !facing), 3);
+
+            te.setIsFacingForwards(tileData, !te.getIsFacingForwards(tileData));
+            be.setChanged();
+
             player.displayClientMessage(new TranslatableComponent("extendedbogeys.tooltips.rotation"), true);
             return InteractionResult.CONSUME;
         }
@@ -114,12 +113,14 @@ public abstract class MixinStandardBogeyBlock extends Block {
                 && !player.getCooldowns().isOnCooldown(wrenchItem) && interactionHand == InteractionHand.MAIN_HAND) {
             player.getCooldowns().addCooldown(wrenchItem, 20);
 
-            int bogeyStyle = state.getValue(STYLE);
-            bogeyStyle = bogeyStyle >= BogeyStyles.getNumberOfBogeyStyleVariations() ? 0 : bogeyStyle + 1;
+            int bogeyStyle = te.getBogeyStyle(tileData);
+            bogeyStyle = bogeyStyle >= BogeyStyles.getNumberOfBogeyStyleVariations()-1 ? 0 : bogeyStyle + 1;
 
             IBogeyStyle style = BogeyStyles.getBogeyStyle(bogeyStyle);
 
-            level.setBlock(blockPos, state.setValue(STYLE, bogeyStyle), 3);
+            te.setBogeyStyle(tileData, bogeyStyle);
+            be.setChanged();
+
             player.displayClientMessage(new TextComponent("Bogey Style: " + bogeyStyle + " \"" + style.getStyleName() + "\""), true);
 
             return InteractionResult.CONSUME;
@@ -127,35 +128,36 @@ public abstract class MixinStandardBogeyBlock extends Block {
         return InteractionResult.PASS;
     }
 
-    /**
-     * @author Rabbitminers / Extended Bogeys
-     * @reason Replace rendering with style based system
-     */
     @OnlyIn(Dist.CLIENT)
-    @Overwrite(remap = false)
-    public void render(BlockState state, float wheelAngle, PoseStack ms, float partialTicks, MultiBufferSource buffers,
-                       int light, int overlay) {
-        boolean isFacingForward = true;
-        int style = 0;
+    @Override
+    public void renderWithTileEntity(BlockState state, BlockEntity te, float wheelAngle, PoseStack ms,
+                                     float partialTicks, MultiBufferSource buffers, int light, int overlay) {
         if (state != null) {
             ms.translate(.5f, .5f, .5f);
             if (state.getValue(AXIS) == Direction.Axis.X)
                 ms.mulPose(Vector3f.YP.rotationDegrees(90));
-            isFacingForward = state.getValue(IS_FACING_FORWARD);
-            style = state.getValue(STYLE);
         }
+        if (te == null)
+            return;
+
+        CompoundTag tileData = te.getTileData();
+
+        IStyledStandardBogeyTileEntity ssbte = (IStyledStandardBogeyTileEntity) te;
+
+        boolean isFacingForward = ssbte.getIsFacingForwards(tileData);
+        int style = ssbte.getBogeyStyle(tileData);
 
         ms.translate(0, -1.5 - 1 / 128f, 0);
 
         VertexConsumer vb = buffers.getBuffer(RenderType.cutoutMipped());
-        BlockState air = Blocks.AIR.defaultBlockState();
 
+        BlockState air = Blocks.AIR.defaultBlockState();
         IBogeyStyle bogeyStyle = BogeyStyles.getBogeyStyle(style);
 
         if (bogeyStyle.shouldRenderInnerShaft())
             for (int i : Iterate.zeroAndOne)
                 CachedBufferer.block(AllBlocks.SHAFT.getDefaultState()
-                        .setValue(ShaftBlock.AXIS, Direction.Axis.Z))
+                                .setValue(ShaftBlock.AXIS, Direction.Axis.Z))
                         .translate(-.5f, .25f, i * -1)
                         .centre()
                         .rotateZ(wheelAngle)
